Refactor chart into own file
