+++ src/components/Chloropleth.jsx
@@ -314,6 +314,12 @@ const Chloropleth = (props) => {
             onHover={e => {
               const [feature] = e.features
               if (feature && 'value' in feature.properties) {
+                if (feature.properties.latitude === undefined) {
+                  // Hack where if no central point is specified
+                  // we use mouse position for popup
+                  feature.properties.lat = e.lngLat[1]
+                  feature.properties.long = e.lngLat[0]
+                }
                 setPopupFeature(feature.properties)
               } else {
                 setPopupFeature(null)
